Use close_fds=True for all subprocess invocations
The Python default is to keep all FDs open, which is how Unix
traditionally worked. This default behavior is particularly
problematic for the atomic command, which e.g. may
hold a reference to the Docker socket, and we don't
necessarily want to give that reference to all
child processes.

This patch changes most calls to subprocess to be through
util package.  And the util subprocess calls all close_fds=True.