Run SPC containers with /sys/fs/selinux mounted as read/only
This tells libselinux inside the container that SELinux is disabled,
even though we are in a SPC container, we don't want the container attempting
to do labeling and especially do not want it loading policy.  If I accidently
load a RHEL6 policy on a RHEL7 machine, bad stuff will happen.